Created yaml models for alll alert subtypes. created classes for all new sub alert types.perform query returns sublclass alert type if it exists. moved Legacy setter criterion to new file. Renamed BaseAlert to alert. Still have mixin issues for legacy methods.
