+++ src/cbc_sdk/platform/models/alert_note.yaml
